Make LangUtilsHook a true hook. Create LangUtils util class
