+++ src/main/java/world/bentobox/bentobox/hooks/LangUtilsHook.java
@@ -1,8 +1,6 @@
 package world.bentobox.bentobox.hooks;
 
-import java.util.Locale;
 import java.util.Map.Entry;
-import java.util.logging.Logger;
 
 import org.bukkit.DyeColor;
 import org.bukkit.Material;
@@ -14,9 +12,7 @@ import org.bukkit.entity.EntityType;
 import org.bukkit.entity.TropicalFish;
 import org.bukkit.entity.Villager;
 import org.bukkit.inventory.ItemStack;
-import org.bukkit.inventory.meta.ItemMeta;
 import org.bukkit.inventory.meta.TropicalFishBucketMeta;
-import org.bukkit.plugin.Plugin;
 import org.bukkit.potion.PotionEffect;
 import org.bukkit.potion.PotionEffectType;
 import org.bukkit.potion.PotionType;
@@ -24,10 +20,8 @@ import org.jetbrains.annotations.Nullable;
 
 import com.meowj.langutils.lang.LanguageHelper;
 
-import world.bentobox.bentobox.BentoBox;
 import world.bentobox.bentobox.api.hooks.Hook;
 import world.bentobox.bentobox.api.user.User;
-import world.bentobox.bentobox.util.Util;
 
 /**
  * @author ApacheZy
@@ -35,56 +29,21 @@ import world.bentobox.bentobox.util.Util;
  */
 public class LangUtilsHook extends Hook {
 
-    private static boolean hooked;
-
     public LangUtilsHook() {
         super("LangUtils", Material.BOOK);
     }
 
-    private static boolean doHook(Plugin plugin) {
-
-        // Because there are other plugins with the same name,
-        // we should check here whether it is the plugin we need.
-
-        if (plugin != null && plugin.isEnabled()) {
-
-            String tag = plugin.getConfig().getString("Extra-TAG");
-            if ("tag_r72EhIAL".equals(tag)) {
-                hooked = true;
-                return true;
-            }
-
-            Logger logger = BentoBox.getInstance().getLogger();
-            logger.warning("This LangUtils version is not available for BentoBox.");
-            logger.warning("Please go here to download the latest version:");
-            logger.warning("https://github.com/apachezy/LangUtils/releases");
-        }
-
-        hooked = false;
-        return false;
-    }
-
     @Override
     public boolean hook() {
-        return LangUtilsHook.doHook(getPlugin());
+        return getPlugin() != null && getPlugin().isEnabled() && getPlugin().getConfig().getString("Extra-TAG", "").equals("tag_r72EhIAL");
     }
 
     @Override
     public String getFailureCause() {
-        return "The LangUtils version does not apply to BentoBox.";
+        return "The LangUtils version does not apply to BentoBox. Download the latest version: https://github.com/apachezy/LangUtils/releases";
     }
 
-    /**
-     * Sometimes it is necessary to check whether "LangUtils" exists
-     * first to decide what method to use to complete the work.
-     *
-     * @return LangUtils is loaded correctly.
-     */
-    public static boolean isHooked() {
-        return hooked;
-    }
-
-    private static String getUserLocale(User user) {
+    private String getUserLocale(User user) {
         return user.getLocale().toLanguageTag();
     }
 
@@ -99,18 +58,8 @@ public class LangUtilsHook extends Hook {
      * @param user the User's locale will be used for translation.
      * @return The Display-Name of the item.
      */
-    public static String getItemDisplayName(ItemStack item, User user) {
-        if (hooked) {
-            return LanguageHelper.getItemDisplayName(item, getUserLocale(user));
-        }
-        ItemMeta meta = item.getItemMeta();
-        if (meta != null && meta.hasDisplayName()) {
-            String dname = meta.getDisplayName();
-            if (!dname.isEmpty()) {
-                return dname;
-            }
-        }
-        return Util.prettifyText(item.getType().name());
+    public String getItemDisplayName(ItemStack item, User user) {
+        return LanguageHelper.getItemDisplayName(item, getUserLocale(user));
     }
 
     /**
@@ -123,10 +72,8 @@ public class LangUtilsHook extends Hook {
      * @param user      the User's locale will be used for translation.
      * @return The translated item name.
      */
-    public static String getItemName(ItemStack itemStack, User user) {
-        return hooked
-                ? LanguageHelper.getItemName(itemStack, getUserLocale(user))
-                : Util.prettifyText(itemStack.getType().name());
+    public String getItemName(ItemStack itemStack, User user) {
+        return LanguageHelper.getItemName(itemStack, getUserLocale(user));
     }
 
     /**
@@ -139,23 +86,8 @@ public class LangUtilsHook extends Hook {
      * @param user     the User's locale will be used for translation.
      * @return The translated material name.
      */
-    public static String getMaterialName(Material material, User user) {
-        return hooked
-                ? LanguageHelper.getMaterialName(material, getUserLocale(user))
-                : Util.prettifyText(material.name());
-    }
-
-    /**
-     * Return the display name of the entity.
-     *
-     * @param entity The entity
-     * @param user   the User's locale will be used for translation.
-     * @return The name of the entity
-     */
-    public static String getEntityDisplayName(Entity entity, User user) {
-        return entity.getCustomName() != null
-                ? entity.getCustomName()
-                : getEntityName(entity, user);
+    public String getMaterialName(Material material, User user) {
+        return LanguageHelper.getMaterialName(material, getUserLocale(user));
     }
 
     /**
@@ -165,10 +97,8 @@ public class LangUtilsHook extends Hook {
      * @param user       the User's locale will be used for translation.
      * @return The translated EntityType name.
      */
-    public static String getEntityName(EntityType entityType, User user) {
-        return hooked
-                ? LanguageHelper.getEntityName(entityType, getUserLocale(user))
-                : Util.prettifyText(entityType.toString());
+    public String getEntityName(EntityType entityType, User user) {
+        return LanguageHelper.getEntityName(entityType, getUserLocale(user));
     }
 
     /**
@@ -178,10 +108,8 @@ public class LangUtilsHook extends Hook {
      * @param user   the User's locale will be used for translation.
      * @return The translated EntityType name.
      */
-    public static String getEntityName(Entity entity, User user) {
-        return hooked
-                ? LanguageHelper.getEntityName(entity, getUserLocale(user))
-                : Util.prettifyText(entity.getType().toString());
+    public String getEntityName(Entity entity, User user) {
+        return LanguageHelper.getEntityName(entity, getUserLocale(user));
     }
 
     /**
@@ -191,10 +119,8 @@ public class LangUtilsHook extends Hook {
      * @param user  the User's locale will be used for translation.
      * @return The translated Biome name.
      */
-    public static String getBiomeName(Biome biome, User user) {
-        return hooked
-                ? LanguageHelper.getBiomeName(biome, getUserLocale(user))
-                : Util.prettifyText(biome.name());
+    public String getBiomeName(Biome biome, User user) {
+        return LanguageHelper.getBiomeName(biome, getUserLocale(user));
     }
 
     /**
@@ -205,10 +131,8 @@ public class LangUtilsHook extends Hook {
      * @param user  The User's locale will be used for translation.
      * @return Translated enchanted name with level.
      */
-    public static String getEnchantDisplayName(Enchantment ench, int level, User user) {
-        return hooked
-                ? LanguageHelper.getEnchantmentDisplayName(ench, level, getUserLocale(user))
-                : ench.getKey().getKey() + " " + level;
+    public String getEnchantDisplayName(Enchantment ench, int level, User user) {
+        return LanguageHelper.getEnchantmentDisplayName(ench, level, getUserLocale(user));
     }
 
     /**
@@ -219,10 +143,8 @@ public class LangUtilsHook extends Hook {
      * @param user  The User's locale will be used for translation.
      * @return Translated enchanted name with level.
      */
-    public static String getEnchantDisplayName(Entry<Enchantment, Integer> entry, User user) {
-        return hooked
-                ? LanguageHelper.getEnchantmentDisplayName(entry, getUserLocale(user))
-                : entry.getKey().getKey().getKey() + " " + entry.getValue();
+    public String getEnchantDisplayName(Entry<Enchantment, Integer> entry, User user) {
+        return LanguageHelper.getEnchantmentDisplayName(entry, getUserLocale(user));
     }
 
     /**
@@ -232,10 +154,8 @@ public class LangUtilsHook extends Hook {
      * @param user    The User's locale will be used for translation.
      * @return The translated enchant name.
      */
-    public static String getEnchantName(Enchantment enchant, User user) {
-        return hooked
-                ? LanguageHelper.getEnchantmentName(enchant, getUserLocale(user))
-                : enchant.getKey().getKey();
+    public String getEnchantName(Enchantment enchant, User user) {
+        return LanguageHelper.getEnchantmentName(enchant, getUserLocale(user));
     }
 
     /**
@@ -246,10 +166,8 @@ public class LangUtilsHook extends Hook {
      * @param user  The user's language will be used for translation.
      * @return The converted enchantment level.
      */
-    public static String getEnchantLevelName(int level, User user) {
-        return hooked
-                ? LanguageHelper.getEnchantmentLevelName(level, getUserLocale(user))
-                : String.valueOf(level);
+    public String getEnchantLevelName(int level, User user) {
+        return LanguageHelper.getEnchantmentLevelName(level, getUserLocale(user));
     }
 
     /**
@@ -259,36 +177,8 @@ public class LangUtilsHook extends Hook {
      * @param user       The user's language will be used for translation.
      * @return Translated potion name.
      */
-    public static String getPotionTypeName(PotionType potionType, User user) {
-        if (hooked) {
-            return LanguageHelper.getPotionName(potionType, getUserLocale(user));
-        }
-        switch (potionType) {
-            case UNCRAFTABLE:     return "Uncraftable Potion";
-            case WATER:           return "Water Bottle";
-            case MUNDANE:         return "Mundane Potion";
-            case THICK:           return "Thick Potion";
-            case AWKWARD:         return "Awkward Potion";
-            case NIGHT_VISION:    return "Potion of Night Vision";
-            case INVISIBILITY:    return "Potion of Invisibility";
-            case JUMP:            return "Potion of Leaping";
-            case FIRE_RESISTANCE: return "Potion of Fire Resistance";
-            case SPEED:           return "Potion of Swiftness";
-            case SLOWNESS:        return "Potion of Slowness";
-            case WATER_BREATHING: return "Potion of Water Breathing";
-            case INSTANT_HEAL:    return "Potion of Healing";
-            case INSTANT_DAMAGE:  return "Potion of Harming";
-            case POISON:          return "Potion of Poison";
-            case REGEN:           return "Potion of Regeneration";
-            case STRENGTH:        return "Potion of Strength";
-            case WEAKNESS:        return "Potion of Weakness";
-            case LUCK:            return "Potion of Luck";
-            case TURTLE_MASTER:   return "Potion of the Turtle Master";
-            case SLOW_FALLING:    return "Potion of Slow Falling";
-            default:
-                return Util.prettifyText(potionType.name());
-        }
-
+    public String getPotionTypeName(PotionType potionType, User user) {
+        return LanguageHelper.getPotionName(potionType, getUserLocale(user));
     }
 
     /**
@@ -298,35 +188,9 @@ public class LangUtilsHook extends Hook {
      * @param user       The user's language will be used for translation.
      * @return Translated splash potion name.
      */
-    public static String getSplashPotionName(PotionType potionType, User user) {
-        if (hooked) {
-            return LanguageHelper.getSplashPotionName(potionType, getUserLocale(user));
-        }
-        switch (potionType) {
-            case UNCRAFTABLE:     return "Splash Uncraftable Potion";
-            case WATER:           return "Splash Water Bottle";
-            case MUNDANE:         return "Mundane Splash Potion";
-            case THICK:           return "Thick Splash Potion";
-            case AWKWARD:         return "Awkward Splash Potion";
-            case NIGHT_VISION:    return "Splash Potion of Night Vision";
-            case INVISIBILITY:    return "Splash Potion of Invisibility";
-            case JUMP:            return "Splash Potion of Leaping";
-            case FIRE_RESISTANCE: return "Splash Potion of Fire Resistance";
-            case SPEED:           return "Splash Potion of Swiftness";
-            case SLOWNESS:        return "Splash Potion of Slowness";
-            case WATER_BREATHING: return "Splash Potion of Water Breathing";
-            case INSTANT_HEAL:    return "Splash Potion of Healing";
-            case INSTANT_DAMAGE:  return "Splash Potion of Harming";
-            case POISON:          return "Splash Potion of Poison";
-            case REGEN:           return "Splash Potion of Regeneration";
-            case STRENGTH:        return "Splash Potion of Strength";
-            case WEAKNESS:        return "Splash Potion of Weakness";
-            case LUCK:            return "Splash Potion of Luck";
-            case TURTLE_MASTER:   return "Splash Potion of the Turtle Master";
-            case SLOW_FALLING:    return "Splash Potion of Slow Falling";
-            default:
-                return Util.prettifyText(potionType.name());
-        }
+    public String getSplashPotionName(PotionType potionType, User user) {
+        return LanguageHelper.getSplashPotionName(potionType, getUserLocale(user));
+
     }
 
     /**
@@ -336,35 +200,8 @@ public class LangUtilsHook extends Hook {
      * @param user       The user's language will be used for translation.
      * @return Translated lingering potion name.
      */
-    public static String getLingeringPotionName(PotionType potionType, User user) {
-        if (hooked) {
-            return LanguageHelper.getLingeringPotionName(potionType, getUserLocale(user));
-        }
-        switch (potionType) {
-            case UNCRAFTABLE:     return "Lingering Uncraftable Potion";
-            case WATER:           return "Lingering Water Bottle";
-            case MUNDANE:         return "Mundane Lingering Potion";
-            case THICK:           return "Thick Lingering Potion";
-            case AWKWARD:         return "Awkward Lingering Potion";
-            case NIGHT_VISION:    return "Lingering Potion of Night Vision";
-            case INVISIBILITY:    return "Lingering Potion of Invisibility";
-            case JUMP:            return "Lingering Potion of Leaping";
-            case FIRE_RESISTANCE: return "Lingering Potion of Fire Resistance";
-            case SPEED:           return "Lingering Potion of Swiftness";
-            case SLOWNESS:        return "Lingering Potion of Slowness";
-            case WATER_BREATHING: return "Lingering Potion of Water Breathing";
-            case INSTANT_HEAL:    return "Lingering Potion of Healing";
-            case INSTANT_DAMAGE:  return "Lingering Potion of Harming";
-            case POISON:          return "Lingering Potion of Poison";
-            case REGEN:           return "Lingering Potion of Regeneration";
-            case STRENGTH:        return "Lingering Potion of Strength";
-            case WEAKNESS:        return "Lingering Potion of Weakness";
-            case LUCK:            return "Lingering Potion of Luck";
-            case TURTLE_MASTER:   return "Lingering Potion of the Turtle Master";
-            case SLOW_FALLING:    return "Lingering Potion of Slow Falling";
-            default:
-                return Util.prettifyText(potionType.name());
-        }
+    public String getLingeringPotionName(PotionType potionType, User user) {
+        return LanguageHelper.getLingeringPotionName(potionType, getUserLocale(user));
     }
 
     /**
@@ -374,35 +211,8 @@ public class LangUtilsHook extends Hook {
      * @param user       The user's language will be used for translation.
      * @return Translated tipped arrow name.
      */
-    public static String getTippedArrowName(PotionType potionType, User user) {
-        if (hooked) {
-            return LanguageHelper.getTippedArrowName(potionType, getUserLocale(user));
-        }
-        switch (potionType) {
-            case UNCRAFTABLE:     return "Uncraftable Tipped Arrow";
-            case WATER:           return "Arrow of Splashing";
-            case MUNDANE:
-            case THICK:
-            case AWKWARD:         return "Tipped Arrow";
-            case NIGHT_VISION:    return "Arrow of Night Vision";
-            case INVISIBILITY:    return "Arrow of Invisibility";
-            case JUMP:            return "Arrow of Leaping";
-            case FIRE_RESISTANCE: return "Arrow of Fire Resistance";
-            case SPEED:           return "Arrow of Swiftness";
-            case SLOWNESS:        return "Arrow of Slowness";
-            case WATER_BREATHING: return "Arrow of Water Breathing";
-            case INSTANT_HEAL:    return "Arrow of Healing";
-            case INSTANT_DAMAGE:  return "Arrow of Harming";
-            case POISON:          return "Arrow of Poison";
-            case REGEN:           return "Arrow of Regeneration";
-            case STRENGTH:        return "Arrow of Strength";
-            case WEAKNESS:        return "Arrow of Weakness";
-            case LUCK:            return "Arrow of Luck";
-            case TURTLE_MASTER:   return "Arrow of the Turtle Master";
-            case SLOW_FALLING:    return "Arrow of Slow Falling";
-            default:
-                return Util.prettifyText(potionType.name());
-        }
+    public String getTippedArrowName(PotionType potionType, User user) {
+        return LanguageHelper.getTippedArrowName(potionType, getUserLocale(user));
     }
 
     /**
@@ -414,15 +224,8 @@ public class LangUtilsHook extends Hook {
      * @param user       The user's language will be used for translation.
      * @return Return the translation result.
      */
-    public static String getPotionBaseEffectName(PotionType potionType, User user) {
-        if (hooked) {
-            return LanguageHelper.getPotionBaseEffectName(potionType, getUserLocale(user));
-        }
-        PotionEffectType effectType = potionType.getEffectType();
-        if (effectType == null) {
-            return "No Effects";
-        }
-        return Util.prettifyText(effectType.getName());
+    public String getPotionBaseEffectName(PotionType potionType, User user) {
+        return LanguageHelper.getPotionBaseEffectName(potionType, getUserLocale(user));
     }
 
     /**
@@ -432,10 +235,8 @@ public class LangUtilsHook extends Hook {
      * @param user       The user's language will be used for translation.
      * @return Translated name of potion effect.
      */
-    public static String getPotionEffectName(PotionEffectType effectType, User user) {
-        return hooked
-                ? LanguageHelper.getPotionEffectName(effectType, getUserLocale(user))
-                : Util.prettifyText(effectType.getName());
+    public String getPotionEffectName(PotionEffectType effectType, User user) {
+        return LanguageHelper.getPotionEffectName(effectType, getUserLocale(user));
     }
 
     /**
@@ -445,11 +246,8 @@ public class LangUtilsHook extends Hook {
      * @param user      The user's language will be used for translation.
      * @return The translated name of the potion level.
      */
-    public static String getEffectAmplifierName(int amplifier, User user) {
-        if (hooked) {
-            return LanguageHelper.getEffectAmplifierName(amplifier, getUserLocale(user));
-        }
-        return amplifier > 0 ? Integer.toString(amplifier) : "";
+    public String getEffectAmplifierName(int amplifier, User user) {
+        return LanguageHelper.getEffectAmplifierName(amplifier, getUserLocale(user));
     }
 
     /**
@@ -459,27 +257,8 @@ public class LangUtilsHook extends Hook {
      * @param user   The user's language will be used for translation.
      * @return The translated and formatted potion effect name, level, and duration.
      */
-    public static String getPotionEffectDisplay(PotionEffect effect, User user) {
-        if (hooked) {
-            return LanguageHelper.getPotionEffectDisplay(effect, getUserLocale(user));
-        }
-
-        String effecName = getPotionEffectName(effect.getType(), user);
-        String amplifier = getEffectAmplifierName(effect.getAmplifier(), user);
-
-        if (amplifier.length() > 0) {
-            effecName = effecName + " " + amplifier;
-        }
-
-        int duration = effect.getDuration();
-        if (duration > 20) {
-            int m = duration / 20 / 60;
-            int s = duration / 20 % 60;
-            String ts = String.format("%d:%02d", m, s);
-            effecName = String.format("%s (%s)", effecName, ts);
-        }
-
-        return effecName;
+    public String getPotionEffectDisplay(PotionEffect effect, User user) {
+        return LanguageHelper.getPotionEffectDisplay(effect, getUserLocale(user));
     }
 
     /**
@@ -489,10 +268,8 @@ public class LangUtilsHook extends Hook {
      * @param user        The user's language will be used for translation.
      * @return The translated name of the tropical fish type.
      */
-    public static String getTropicalFishTypeName(TropicalFish.Pattern fishPattern, User user) {
-        return hooked
-                ? LanguageHelper.getTropicalFishTypeName(fishPattern, getUserLocale(user))
-                : Util.prettifyText(fishPattern.name());
+    public String getTropicalFishTypeName(TropicalFish.Pattern fishPattern, User user) {
+        return LanguageHelper.getTropicalFishTypeName(fishPattern, getUserLocale(user));
     }
 
     /**
@@ -506,50 +283,8 @@ public class LangUtilsHook extends Hook {
      *         tropical fish, otherwise return null.
      */
     @Nullable
-    public static String getPredefinedTropicalFishName(TropicalFishBucketMeta meta, User user) {
-        if (hooked) {
-            return LanguageHelper.getPredefinedTropicalFishName(meta, getUserLocale(user));
-        }
-
-        if (meta.hasVariant()) {
-            TropicalFish.Pattern pattern = meta.getPattern();
-
-            // https://minecraft.gamepedia.com/Tropical_Fish#Entity_data
-
-            int type = pattern.ordinal() > 5 ? 1 : 0;
-            int patt = pattern.ordinal() % 6;
-            int bcol = meta.getBodyColor().ordinal();
-            int pcol = meta.getPatternColor().ordinal();
-
-            int variant = (pcol & 255) << 24 | (bcol & 255) << 16 | (patt & 255) << 8 | type;
-
-            switch (variant) {
-                case 117506305: return "Anemone";
-                case 117899265: return "Black Tang";
-                case 185008129: return "Blue Tang";
-                case 117441793: return "Butterflyfish";
-                case 118161664: return "Cichlid";
-                case 65536    : return "Clownfish";
-                case 50726144 : return "Cotton Candy Betta";
-                case 67764993 : return "Dottyback";
-                case 234882305: return "Emperor Red Snapper";
-                case 67110144 : return "Goatfish";
-                case 117441025: return "Moorish Idol";
-                case 16778497 : return "Ornate Butterflyfish";
-                case 101253888: return "Parrotfish";
-                case 50660352 : return "Queen Angelfish";
-                case 918529   : return "Red Cichlid";
-                case 235340288: return "Red Lipped Blenny";
-                case 918273   : return "Red Snapper";
-                case 67108865 : return "Threadfin";
-                case 917504   : return "Tomato Clownfish";
-                case 459008   : return "Triggerfish";
-                case 67699456 : return "Yellowtail Parrotfish";
-                case 67371009 : return "Yellow Tang";
-                default       : break;
-            }
-        }
-        return null;
+    public String getPredefinedTropicalFishName(TropicalFishBucketMeta meta, User user) {
+        return LanguageHelper.getPredefinedTropicalFishName(meta, getUserLocale(user));
     }
 
     /**
@@ -559,10 +294,8 @@ public class LangUtilsHook extends Hook {
      * @param user  The user's language will be used for translation.
      * @return The name of the dye color that has been translated.
      */
-    public static String getDyeColorName(DyeColor color, User user) {
-        return hooked
-                ? LanguageHelper.getDyeColorName(color, getUserLocale(user))
-                : Util.prettifyText(color.name());
+    public String getDyeColorName(DyeColor color, User user) {
+        return LanguageHelper.getDyeColorName(color, getUserLocale(user));
     }
 
     /**
@@ -572,10 +305,8 @@ public class LangUtilsHook extends Hook {
      * @param user  The user's language will be used for translation.
      * @return Translated name of merchant's level.
      */
-    public static String getVillagerLevelName(int level, User user) {
-        return hooked
-                ? LanguageHelper.getVillagerLevelName(level, getUserLocale(user))
-                : Integer.toString(level);
+    public String getVillagerLevelName(int level, User user) {
+        return LanguageHelper.getVillagerLevelName(level, getUserLocale(user));
     }
 
     /**
@@ -585,10 +316,8 @@ public class LangUtilsHook extends Hook {
      * @param user       The user's language will be used for translation.
      * @return The translated profession name of the villager.
      */
-    public static String getVillagerProfessionName(Villager.Profession profession, User user) {
-        return hooked
-                ? LanguageHelper.getVillagerProfessionName(profession, getUserLocale(user))
-                : Util.prettifyText(profession.name());
+    public String getVillagerProfessionName(Villager.Profession profession, User user) {
+        return LanguageHelper.getVillagerProfessionName(profession, getUserLocale(user));
     }
 
     /**
@@ -598,45 +327,8 @@ public class LangUtilsHook extends Hook {
      * @param user    The user's language will be used for translation.
      * @return The translated name of banner pattern.
      */
-    public static String getBannerPatternName(Pattern pattern, User user) {
-        return hooked
-                ? LanguageHelper.getBannerPatternName(pattern, getUserLocale(user))
-                : pattern.getColor().name().toLowerCase(Locale.ROOT)
-                + "_"
-                + pattern.getPattern().name().toLowerCase(Locale.ROOT);
-    }
-
-    /**
-     * Get the description of the music disk.
-     *
-     * @param material Material for music records.
-     * @param user     This parameter is temporarily useless and is reserved for
-     *                 possible future changes.
-     * @return If the given material is a music disk, the description of the
-     *         music disk is returned. Otherwise, return NULL.
-     */
-    @Nullable
-    public static String getMusicDiskDesc(Material material, User user) {
-
-        // The description of the music record is the same in any language,
-        // so directly output it here.
-
-        switch (material) {
-            case MUSIC_DISC_13      : return "C418 - 13";
-            case MUSIC_DISC_CAT     : return "C418 - cat";
-            case MUSIC_DISC_BLOCKS  : return "C418 - blocks";
-            case MUSIC_DISC_CHIRP   : return "C418 - chirp";
-            case MUSIC_DISC_FAR     : return "C418 - far";
-            case MUSIC_DISC_MALL    : return "C418 - mall";
-            case MUSIC_DISC_MELLOHI : return "C418 - mellohi";
-            case MUSIC_DISC_STAL    : return "C418 - stal";
-            case MUSIC_DISC_STRAD   : return "C418 - strad";
-            case MUSIC_DISC_WARD    : return "C418 - ward";
-            case MUSIC_DISC_11      : return "C418 - 11";
-            case MUSIC_DISC_WAIT    : return "C418 - wait";
-            case MUSIC_DISC_PIGSTEP : return "Lena Raine - Pigstep";
-            default                 : return null;
-        }
+    public String getBannerPatternName(Pattern pattern, User user) {
+        return LanguageHelper.getBannerPatternName(pattern, getUserLocale(user));
     }
 
 }
