+++ src/main/java/world/bentobox/bentobox/listeners/flags/clicklisteners/GeoMobLimitTab.java
@@ -13,7 +13,6 @@ import org.bukkit.event.inventory.ClickType;
 import org.eclipse.jdt.annotation.NonNull;
 import org.eclipse.jdt.annotation.Nullable;
 
-
 import world.bentobox.bentobox.BentoBox;
 import world.bentobox.bentobox.api.addons.GameModeAddon;
 import world.bentobox.bentobox.api.panels.Panel;
@@ -23,7 +22,7 @@ import world.bentobox.bentobox.api.panels.Tab;
 import world.bentobox.bentobox.api.panels.TabbedPanel;
 import world.bentobox.bentobox.api.panels.builders.PanelItemBuilder;
 import world.bentobox.bentobox.api.user.User;
-import world.bentobox.bentobox.hooks.LangUtilsHook;
+import world.bentobox.bentobox.util.LangUtils;
 import world.bentobox.bentobox.util.Util;
 
 /**
@@ -121,7 +120,7 @@ public class GeoMobLimitTab implements Tab, ClickHandler {
 
     private PanelItem getPanelItem(EntityType c, User user) {
         PanelItemBuilder pib = new PanelItemBuilder();
-        pib.name(LangUtilsHook.getEntityName(c, user));
+        pib.name(LangUtils.getEntityName(c, user));
         pib.clickHandler(this);
         if (type == EntityLimitTabType.MOB_LIMIT) {
             if (!BentoBox.getInstance().getIWM().getMobLimitSettings(world).contains(c.name())) {
